feat: next-vercel function interoperability
