+++ frontend/tsconfig.json
@@ -30,5 +30,5 @@
     "strictNullChecks": false
   },
   "exclude": ["node_modules"],
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"]
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "api/newsletter-signup.mjs"]
 }
